// src/app/page/master/deskripsi/deskripsi.interface.ts
import {EventEmitter} from '@angular/core';

export interface Deskripsi {
	namaDeskripsi?: any;
	detailDeskripsi?: any;
     persenHargaCito?: any;
	 namaJenisDeskripsi?: any;    
     namaKategoryDeskripsi?: any;
     noUrut?: any;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}